+++ sites/admin/assets/css/choose-witch.css
