+++ sites/admin/design/contexts/standard.php
@@ -27,13 +27,13 @@
         <!-- content -->
         <section>
             <div class="breadcrumb">
-                <span class="breadcrumb__label">
-                    Breadcrumb&nbsp;:
-                </span>
+                <h1 class="breadcrumb__label" title="<?=$this->wc->witch()->data ?>">
+                    <a href="javascript: location.reload();"><?=$this->wc->witch()->name ?></a>
+                </h1>
                 
                 <?php foreach( $breadcrumb as $i => $breadcrumbItem ): ?>
                     <?=( $i > 0 )? "&nbsp;>&nbsp": "" ?>
-                    <span class="breadcrumb__item">
+                    <span class="breadcrumb__item" title="<?=$breadcrumbItem['data'] ?>">
                         <a href="<?=$breadcrumbItem['href'] ?>">
                             <?=$breadcrumbItem['name'] ?>
                         </a>
@@ -46,12 +46,23 @@
                 <?php if( $this->wc->witch("arborescence") ): ?>
                     <a class="tabs__item" href="#tab-navigation">
                         <i class="fas fa-sitemap"></i>
+                        <!--i class="fas fa-broom"></i-->
+                        Navigation
                     </a>
                 <?php endif; ?>
-                <a class="tabs__item selected" href="#tab-current">                   
-                    <?php if( !$this->wc->witch() ): ?>
+                
+                <?php if( !$this->wc->witch() ): ?>
+                    <a class="tabs__item selected" href="#tab-current">                   
                         404
-                    <?php else: ?>
+                    </a>
+                <?php elseif( $this->tabs ): foreach( $this->tabs as $id => $tab ): ?>
+                    <a class="tabs__item <?=($tab['selected'] ?? null)? 'selected': '' ?>" 
+                       href="#<?=$id ?>">
+                        <?=($tab['iconClass'] ?? null)? '<i  class="'.$tab['iconClass'].'"></i>': '' ?>
+                        <?=$tab['text'] ?? '' ?>
+                    </a>
+                <?php endforeach; else: ?>
+                    <a class="tabs__item selected" href="#tab-current">
                         <?php if( $this->wc->witch()->hasCraft() && $this->wc->witch()->invoke ): ?>
                             <i  class="fas fa-hat-wizard"></i>
                         <?php elseif( $this->wc->witch()->hasCraft() ): ?>
@@ -61,37 +72,41 @@
                         <?php else: ?>
                             <i class="fas fa-folder"></i>
                         <?php endif; ?>
-                        
+
                         <?=$this->wc->witch()->name ?>
-                        
+
                         <?php if( $this->wc->witch("target")->exist() ): ?>
                             &nbsp;:
                         <?php elseif( $this->wc->witch("mother")->exist() ): ?>
                             from&nbsp;:
                         <?php endif; ?>
                         <?=$this->wc->witch("target").$this->wc->witch("mother")?>
-                    <?php endif; ?>
-                </a>
-                <div class="clear"></div>
+                    </a>
+                <?php endif; ?>
             </div>
             
             <div class="tabs-target">
-                <div class="tabs-target__item selected" id="tab-current">
-                    <?php if( !$this->wc->witch() ): ?>
-                        404
-                    <?php else: ?>
-                        <?=$this->wc->witch()->result() ?>
-                    <?php endif; ?>
-                </div>
+                <?php if( !$this->wc->witch() ): ?>
+                    <div class="tabs-target__item selected" id="tab-current">404</div>
+                <?php elseif( !$this->tabs ): ?>
+                    <div class="tabs-target__item selected" id="tab-current"><?=$this->wc->cairn->invokation() ?></div>
+                <?php else: ?>
+                    <?=$this->wc->cairn->invokation() ?>
+                <?php endif; ?>
+                
                 <?php if( $this->wc->witch("arborescence") ): ?>
                     <div class="tabs-target__item" id="tab-navigation">
-                        <?=$this->wc->witch("arborescence")->result("arborescence") ?>
+                        <?=$this->wc->cairn->invokation("arborescence") ?>
                     </div>
                 <?php endif; ?>
             </div>
         </section>
         
+        <?php if( $this->wc->witch("chooseWitch") ): ?>
+            <?=$this->wc->cairn->invokation("chooseWitch") ?>
+        <?php endif; ?>
+        
         <!-- footer -->
-        <?php include $this->getIncludeDesignFile('footer.php'); ?>                
+        <?php include $this->getIncludeDesignFile('footer.php'); ?>
     </body>
 </html>
